Swapped out menu for manage skills with the newer version
Also changed some keycodes for options in the menu to be less about
anchoring to the first letter of words, and more about keeping them
ergonomical and on the left half of the keyboard. Note, might add a
Qud option in the future to swap keycodes for left/right handed.

Renamed some menu options just for visual appeal.