Reverted Equipment/Equipment-Type Preference Functionality
These just weren't working out. With the way reequip works in native
code, it's pretty hacky to try to control how your companion reequips
their character. Creating a solution for this type of behavior is likely
going to be a solution much better suited for the game devs to solve
in a generic way anyways, instead of Clever Girl doing its' own thing.